+++ aries_cloudagent/config/default_context.py
@@ -142,7 +142,6 @@ class DefaultContextBuilder(ContextBuilder):
         plugin_registry.register_plugin("aries_cloudagent.resolver")
         plugin_registry.register_plugin("aries_cloudagent.settings")
         plugin_registry.register_plugin("aries_cloudagent.vc")
-        plugin_registry.register_plugin("aries_cloudagent.vc_api")
         plugin_registry.register_plugin("aries_cloudagent.wallet")
         if wallet_type == "askar-anoncreds":
             plugin_registry.register_plugin("aries_cloudagent.anoncreds")
